requirements.txt
mysql-connector-python==9.0.0
python-dateutil==2.9.0.post0
